fix pg + split mysql & pg builder
