+++ query-engine/connectors/sql-query-connector/src/query_builder/select/mysql.rs
@@ -1,0 +1,229 @@
+use super::*;
+use crate::{
+    context::Context,
+    filter::alias::{Alias, AliasMode},
+    model_extensions::*,
+};
+
+use quaint::ast::*;
+use query_structure::*;
+use std::borrow::Cow;
+
+#[derive(Debug, Default)]
+pub(crate) struct MysqlSelectBuilder {
+    alias: Alias,
+}
+
+impl JoinSelectBuilder for MysqlSelectBuilder {
+    fn build(&mut self, args: QueryArguments, selected_fields: &FieldSelection, ctx: &Context<'_>) -> Select<'static> {
+        let (select, alias) = self.build_default_select(&args, ctx);
+
+        self.with_selection(select, selected_fields, alias, ctx)
+    }
+
+    fn with_selection<'a>(
+        &mut self,
+        select: Select<'a>,
+        selected_fields: &FieldSelection,
+        table_alias: Alias,
+        ctx: &Context<'_>,
+    ) -> Select<'a> {
+        selected_fields
+            .selections()
+            .fold(select, |acc, selection| match selection {
+                SelectedField::Scalar(sf) => acc.column(
+                    sf.as_column(ctx)
+                        .table(table_alias.to_table_string())
+                        .set_is_selected(true),
+                ),
+                SelectedField::Relation(rs) => self.with_relation(acc, rs, table_alias, ctx),
+                _ => acc,
+            })
+    }
+
+    fn with_to_one_relation<'a>(
+        &mut self,
+        select: Select<'a>,
+        rs: &RelationSelection,
+        parent_alias: Alias,
+        ctx: &Context<'_>,
+    ) -> Select<'a> {
+        let (subselect, _) = self.build_to_one_select(rs, parent_alias, |x| x, ctx);
+
+        select.value(Expression::from(subselect).alias(rs.field.name().to_owned()))
+    }
+
+    fn with_to_many_relation<'a>(
+        &mut self,
+        select: Select<'a>,
+        rs: &query_structure::prelude::RelationSelection,
+        parent_alias: Alias,
+        ctx: &Context<'_>,
+    ) -> Select<'a> {
+        let join_table =
+            Expression::from(self.build_to_many_select(rs, parent_alias, ctx)).alias(rs.field.name().to_owned());
+
+        // correlated subquery
+        select.value(join_table)
+    }
+
+    fn with_many_to_many_relation<'a>(
+        &mut self,
+        select: Select<'a>,
+        rs: &query_structure::prelude::RelationSelection,
+        parent_alias: Alias,
+        ctx: &Context<'_>,
+    ) -> Select<'a> {
+        let m2m_select = self.build_m2m_select(rs, parent_alias, ctx);
+
+        select.value(Expression::from(m2m_select).alias(rs.field.name().to_owned()))
+    }
+
+    fn build_json_obj_fn(
+        &mut self,
+        rs: &RelationSelection,
+        ctx: &Context<'_>,
+        parent_alias: Alias,
+    ) -> Expression<'static> {
+        let build_obj_params = rs
+            .selections
+            .iter()
+            .filter_map(|f| match f {
+                SelectedField::Scalar(sf) => Some((
+                    Cow::from(sf.db_name().to_owned()),
+                    Expression::from(sf.as_column(ctx).table(parent_alias.to_table_string())),
+                )),
+                SelectedField::Relation(rs) => Some((
+                    Cow::from(rs.field.name().to_owned()),
+                    Expression::from(self.with_relation(Select::default(), rs, parent_alias, ctx)),
+                )),
+                _ => None,
+            })
+            .collect();
+
+        json_build_object(build_obj_params).into()
+    }
+
+    fn next_alias(&mut self) -> Alias {
+        self.alias = self.alias.inc(AliasMode::Table);
+        self.alias
+    }
+}
+
+impl MysqlSelectBuilder {
+    fn build_to_many_select(
+        &mut self,
+        rs: &RelationSelection,
+        parent_alias: Alias,
+        ctx: &Context<'_>,
+    ) -> Select<'static> {
+        let inner_root_table_alias = self.next_alias();
+        let root_alias = self.next_alias();
+        let inner_alias = self.next_alias();
+        let middle_alias = self.next_alias();
+
+        let related_table = rs
+            .related_model()
+            .as_table(ctx)
+            .alias(inner_root_table_alias.to_table_string());
+
+        // SELECT * FROM "Table" as <table_alias> WHERE parent.id = child.parent_id
+        let root = Select::from_table(related_table)
+            .with_join_conditions(&rs.field, parent_alias, inner_root_table_alias, ctx)
+            .comment("root select");
+
+        // SELECT JSON_BUILD_OBJECT() FROM ( <root> )
+        let inner = Select::from_table(Table::from(root).alias(root_alias.to_table_string()))
+            .value(self.build_json_obj_fn(rs, ctx, root_alias).alias(JSON_AGG_IDENT));
+
+        // LEFT JOIN LATERAL () AS <inner_alias> ON TRUE
+        let inner = self.with_relations(inner, rs.relations(), root_alias, ctx);
+
+        let linking_fields = rs.field.related_field().linking_fields();
+
+        if rs.field.relation().is_many_to_many() {
+            let selection: Vec<Column<'_>> =
+                FieldSelection::union(vec![order_by_selection(rs), linking_fields, filtering_selection(rs)])
+                    .into_projection()
+                    .as_columns(ctx)
+                    .map(|c| c.table(root_alias.to_table_string()))
+                    .collect();
+
+            // SELECT <foreign_keys>, <orderby columns>
+            inner.with_columns(selection.into())
+        } else {
+            // select ordering, filtering & join fields from child selections to order, filter & join them on the outer query
+            let inner_selection: Vec<Column<'_>> = FieldSelection::union(vec![
+                order_by_selection(rs),
+                filtering_selection(rs),
+                relation_selection(rs),
+            ])
+            .into_projection()
+            .as_columns(ctx)
+            .map(|c| c.table(root_alias.to_table_string()))
+            .collect();
+
+            let inner = inner.with_columns(inner_selection.into()).comment("inner select");
+
+            // On MySQL, using LIMIT makes the ordering of the JSON_AGG working. Beware, this is undocumented behavior.
+            let take = rs.args.take_abs().or(Some(i64::MAX));
+
+            let middle = Select::from_table(Table::from(inner).alias(inner_alias.to_table_string()))
+                // SELECT <inner_alias>.<JSON_ADD_IDENT>
+                .column(Column::from((inner_alias.to_table_string(), JSON_AGG_IDENT)))
+                // ORDER BY ...
+                .with_ordering(&rs.args, Some(inner_alias.to_table_string()), ctx)
+                // WHERE ...
+                .with_filters(rs.args.filter.clone(), Some(inner_alias), ctx)
+                // LIMIT $1 OFFSET $2
+                .with_pagination(take, rs.args.skip)
+                .comment("middle select");
+
+            // SELECT COALESCE(JSON_AGG(<inner_alias>), '[]') AS <inner_alias> FROM ( <middle> ) as <inner_alias_2>
+            Select::from_table(Table::from(middle).alias(middle_alias.to_table_string()))
+                .value(json_agg())
+                .comment("outer select")
+        }
+    }
+
+    fn build_m2m_select<'a>(&mut self, rs: &RelationSelection, parent_alias: Alias, ctx: &Context<'_>) -> Select<'a> {
+        let rf = rs.field.clone();
+        let m2m_table_alias = self.next_alias();
+        let root_alias = self.next_alias();
+        let outer_alias = self.next_alias();
+
+        let m2m_join_data =
+            rf.related_model()
+                .as_table(ctx)
+                .on(rf.m2m_join_conditions(Some(m2m_table_alias), None, ctx));
+
+        let m2m_table = rf.as_table(ctx).alias(m2m_table_alias.to_table_string());
+
+        let root = Select::from_table(m2m_table)
+            .inner_join(m2m_join_data)
+            .value(rf.related_model().as_table(ctx).asterisk())
+            .with_ordering(&rs.args, None, ctx) // adds ordering stmts
+            // Keep join conditions _before_ user filters to ensure index is used first
+            .and_where(
+                rf.related_field()
+                    .m2m_join_conditions(Some(m2m_table_alias), Some(parent_alias), ctx),
+            ) // adds join condition to the child table
+            .with_filters(rs.args.filter.clone(), None, ctx) // adds query filters
+            .comment("root");
+
+        // On MySQL, using LIMIT makes the ordering of the JSON_AGG working. Beware, this is undocumented behavior.
+        let take = match rs.args.order_by.is_empty() {
+            true => rs.args.take_abs(),
+            false => rs.args.take_abs().or(Some(i64::MAX)),
+        };
+
+        let inner = Select::from_table(Table::from(root).alias(root_alias.to_table_string()))
+            .value(self.build_json_obj_fn(rs, ctx, root_alias).alias(JSON_AGG_IDENT))
+            .with_pagination(take, rs.args.skip)
+            .comment("inner"); // adds pagination
+
+        Select::from_table(Table::from(inner).alias(outer_alias.to_table_string()))
+            .value(json_agg())
+            .comment("outer")
+    }
+}
